moved App component into components directory
